Handle anisotropy in the PBR shader
Add two options to vtkProperty, and link them to OSPray:

* anisotropy (between 0 and 1) that controls the strength of the anisotropy.
It will be possible to use a texture to hold this value (WIP).

* rotation (between 0 and 1) that rotate the direction of the anisotropy
(ie. the tangent) around the normal counter-clockwise.
A value of 1 correspond to a rotation of 2*PI. If needed, the rotation is done is
the fragment shader because it allows the user to pass the rotaion as a
texture value.