Add outline glow render pass
Outline glow pass draws a glowing outline of a scene. Combined with layered
renderers this can be used to highlight parts of a scene. Outline glow uses
render passes to first draw the scene, apply gaussian blur and then draw the
blurred image only where no pixels were drawn in the original scene. The result
is an outline with a hard edge on the inside and a smooth edge on the outside.