Fixed #17832: Observe NSView frame changes even for NSViews VTK didn't create
Usually VTK creates its own NSViews, and in that case we were already observing NSViewFrameDidChangeNotification and reacting appropriately.

But VTK can also be used in a different way, were clients create their own NSViews and point VTK to them. The SimpleCocoaVTK example does this.  In this case, VTK was not observing for NSVIew frame changes. That was ok unitl b75bc715.  But now that VTK creates its own framebuffer, it really needs to know when the NSView frame changes size.  So now we observe frame changes even for NSViews we didn't create.  In response, GetSize/SetSize will be invoked and do some necessary work. They will not however change the size of the NSView/NSWindow becuase a new 'if' contition limits that to happening only for NSViews created by VTK.