Correct the number of triangles written to binary STL files
The number of triangle written to byte 80 in binary STL files was not
correct if triangle strips or polygons were included in the input.

Updated test to check that the right number of triangles was written
to binary files when triangle strips and polygons are present in the
input.

(cherry picked from commit 24b1934714290a831426c981e026986d4d3e19e1)