Simplify the OpenGL render pocess a bit
VTK will always render into a RenderFramebuffer If Swapbuffers is on it
will also blit into DisplayFramebuffers which can then be used to blit
to and draw destination. The render framebuffer may be multisampled. The
display framebuffers are not so they could be considered resolve
buffers.

Helper functions for GUI integration have been added to blt into the
default hardware buffers or to blit into the current draw destination.

Add a ivar to control frame blit behavior. Defaults to blit to hardware,
but also has blitToCurrent and NoBlit options.