Update VTK-m probe filter for invalid values
The VTK-m probe filter by default now writes NaN into field values of
the sample grid that are outside of the volume being sampled. However,
VTK general expects the value to be set to 0 and to use a separate array
to determine what is really valid. Update the filter to do this
correctly.

Also made a couple of fixes to the VTKMProbe test. It no longer uses
`assert` to test values since this does not actually do anything on
release build tests. It also reconfigures the condition of tests values
to be more stable. It correctly handles the difference regardless of
whether the measured or expected value is low. It also tests the
negative so that if an unexpected NaN is computed, it still fails the
test.