Fix compilation when Qt is built for GLES platforms
On GLES 2.0/3.0 platforms (more specifically, for Qt5 "opengl es2" builds),
QOpenGLFunctions_3_2_Core does not exist.

After the last restructuring, from the GL 3.2 Core functions only
glDrawBuffer is used. glDrawBuffer can be trivially replaced with
glDrawBuffers, which is part of OpenGL 2.0 and GLES 3.0, and as an
extension in many GLES 2.0 implementations.