Rename some OpenGL methods add some docs
Add a convenience method to copy the current data
from the read framebuffer into VTK's render framebuffer

Fix an issue related to resolving MSAA buffers where for
transparent objects the partial occupancy cells were not
computed corrected as they need to be averaged in gamma space.

Add a test for some of the new blitting methods.