Read independent vector components into a single vector array
Additionally, switch to vtkFloatArray from
vtkSOADataArrayTemplate<float> for storing
points. vtkSOADataArrayTemplate isn't well supported in a number of
places in VTK, but particularly in vtkDataWriter, where numeric arrays
are statically cast to vtkDataArray subclasses regardless of their
type. Using vtkFloatArray requires temporary storage of each point or
vector component.  This adds a small cost to the memory footprint of
the reader, basically a single scalar array worth of memory used as a
buffer for vector components.