Fix Issue#17949
Uploading MCWCMatrix (scaled) to the shader to allow the user to find the
coordinates in real space of a point (useful for out of bounds shading for
example). This is difficult to do in the callback because the actor is not
passed to the callback. It is also difficult to put in the vtkShaderProperty on
the actor because it's difficult to know when it's been updated except in the
pipeline

Also uploading the normal transformation matrix due to the expensive cost of
calculating on the shader for every vertex (if needed) as it's an inverse