Restore multiple RenderWindowInteractor functionality.
Since VTK9, on Linux, vtkXRenderWindowInteractor is not capable of
handling having multiple RenderWindowInteractor, while this functionality
was supported in VTK8.2 and is still working on Windows (VTK9).
This commit restores the ability to manage several instances by storing
a pointer to each declared RenderWindowInteractor in a static vector, and
dispatching the events to the right Interactor.