Add support for camera based shift scale methods
Some datasets are so large that no normalization allows them to be
properly represented in float. These datasets can still be renderered
properly if we normalize them based on the current camera/actor pose so
that the visible portion of the dataset is within the normal range of
single precision floats.