stop both mode from blending backplate and environment
ospray 2 will blend the backgrounds if both are enabled. hide the
visibility of the environmental hdri light to prevent that.

meanwhile stop using the background(color) and use
background(texture) instead for ray cast/scivis because it is
now supported. This lets us access gradient and texture backplates
for scivis at long last.