remove a completely extraneous deep copy
This useless code was an artifact of the big messy refactor I did when
hurriedly commited the Extended memory changes to master a few months
back. The code functions fine in either case but as initially committed
we deepcopy twice. We only need the second one and removing this one
speeds things up by a factor of two.