Replaced vtkCriticalSection, vtkMutexLock, vtkConditionVariable with STL
Replaced usage of vtkCriticalSection, vtkMutexLock, and vtkSimpleMutexLock with std::mutex.

Relpaced usage of vtkConditionVariable ad vtkSimpleConditionVariable with std::condition_variable_any.

Marked the those VTK classes as deprecated.  Also deprecated vtkThreadMessanger, which was already unused.

Biggest change is to vtkBlueObeliskData, which publicly exposes a vtkSimpleMutexLock as part of its public API with Get/SetWriteMutex. Made it deprecated and replacing it with new public LockWriteMutex() and UnlockWriteMutex() methods.