Fix couple issues with hardware picking
When interfacing with other OpenGL code we have
to be careful of what we do to the OpenGL state.
The Render() call handles this but there are other
places where user code can cause VTK to execute OpenGL
calls such as picking and reading pixel data without going
through Render().  This topic tries to address the picking
entry point by saving state and restoring it after picking.
It also removes an extra clear that was happening before the
Render() call which was clearing the currently bound buffer
which was unknown and not needed.

Addresses
https://gitlab.kitware.com/paraview/paraview/-/issues/19964