Ensure FBOs are the right size prior to blitting to them
It was possible to call vtkOpenGLRenderWindow::BlitToRenderFramebuffer()
for a render window whose size has changed while the internal FBOs
had the previous size of the render window. This could result in a failure
to blit if the new render window size was larger than the FBO objects.
Fix that by ensuring the framebuffers are resized in advance.