Added a new filter to mark boundary cells, faces, and points
This filter marks cells that are on the boundary, as well
as associated faces and points. A n-D boundary cell is a cell
with a (n-1)-boundary face. A boundary face is used by only
one cell. The filter can be used to compute ghost cells, and
perform other topological analysis of a dataset. The filter
treats any type of vtkDataSet.