Fix point picking with render points as spheres
Addresses issue https://gitlab.kitware.com/vtk/vtk/-/issues/17223

The issue was that when points were rendered as spheres
during point picking. The zbuffer could be initialized with
the spheres. Then in the next pass the coincident code is supposed
to move the zbuffer values a bit closer to make sure they show up
but the zbuffer calc code in render as spheres was overriding
the coincident code (all the time actually)

This change makes it so that the coincident code still has
an impact even with render points as spheres.