Changing output arrays in XML readers
This commit takes advantage of the fact that array element sizes
in XML files are encoded with they lengths in bits. For instance,
instead of instantiating a `vtkLongArray` or a `vtkLongLongArray` for a `Int64`
depending on the compiler and machine building VTK, the reader will
now instantiate a `vtkTypeInt64Array`, which inherits from
`vtkLongArray` or `vtkLongLongArray` depending on the integer encoding
of the machine, at compile time.
The same logic applies for all floating numbers and
integer lengths.