Deflate sampling bounds by epsilon when UseInputBounds is on
To avoid sampling outside the computed bounding box of the input data,
which can happen due to floating point errors in some cases, deflate
the bounding box by epsilon to ensure sampling points close to the
edge of the bounds fall within the dataset. This helps when resampling
vtkImageData and vtkRectilinearGrid datasets where numerical error can
result in a whole planar slab of sample points falling just outside
the input bounding box.

Updated the test to reflect the fact that some points are now found to
be classified differently based on the slightly different computed
volume.