Fixed vtkBoundingBox.GetBounds() when NaNs are present
This commit fixes regression in bounding box computation in vtkBoundingBox class
introduced by Kitware@933371b.
See bug report from user here: https://discourse.vtk.org/t/getbounds-fails-with-nan-points-in-vtk9/4351.

A variable that may contain NaN must not be used as a first argument in comparison,
ternary, or min/max operations, because the result will be NaN. Instead, such variables must
be used as the second second argument. Fix is similar to this commit:
Kitware@06d553f

Full credit goes to @lassoan for this fix