vtkModule: use interface targets to transport private links for kits
When a module private needs an external dependency in a kit build, the
module itself needs the build interface and the kit needs the link
interface. To keep it as private as possible, the kit should link to it
privately, so it should not be part of the module's public interface.
The previous solution to this used a global property to transport the
private targets up and promoted imported targets into global targets.
This is an abuse of the global import mechanism.

Instead, create an interface target with `$<LINK_ONLY:>` usages of the
private libraries and then privately link to this target from the kit if
necessary. This keeps the imported targets properly scoped while still
allowing the kit to link to them exclusively privately (as there is not
"public for me, but unconditionally private for thee" usage requirement
in CMake).

Fixes: #18004