Fix SpecialColors for vtkWindowLevelLookupTable
When the AboveRangeColor and BelowRangeColor were added to the
vtkLookupTable several years ago, the vtkWindowLevelLookupTable
was not similarly modified.  As a result, the values of the
beyond-range colors were left uninitialized and generally had
a value of (0,0,0,0) when the vtkWindowLevelLookupTable was used.

I also took the opportunity to have vtkWindowLevelLookupTable
implement the ForceBuild() method instead of the Build() method.
This allows ForceBuild() and Build() to work for this class in
the same way as they do for vtkLookupTable.