Layer-backed macOS applications shouldn't override the rendering context.
There's been instances with layer-backed macOS applications where contexts (NSOpenGLContext) for views tries to render into the context-object (CGLContextObj) of another, which tends to arise when the context (NSOpenGLContext) gets overridden in the vtkCocoaRenderWindow 'InitializeFromCurrentContext()' function. The idea is to protect the integrity of the context to be rendered into from compromise by falling back onto the 'Superclass::InitializeFromCurrentContext()' instead.