Fix issues with vtkArrayListTemplate use.
320464e6 made `vtkArrayListTemplate` depend on
`vtkDataSetAttributes::RequiredArrays` which means that removing arrays
from output DSA after calls to `InterpolateAllocate` etc. is no longer
valid. When array are excluded, these filters were doing exactly that
causing issues e.g. paraview/paraview#20344.

Fixed by ensuring that before we call InterpolateAllocate, we mark the
excluded arrays as `CopyOff` so they are not allocated in the output DSA
in call to `InterpolateAllocate`.