Simplify coincident topology resolution
This change reverts to an offset in Z-buffer values for coincident topology resolution.
The previous approach specified offsets in view coordinates to account
for the nonlinearity of the Z-buffer. Furthermore, the offset was scaled
using the actor's bounding diagonal. This fixed issues especially with
the surface with edges representation. However, it introduced issues in
application level code where the actor bounding box should not have
been considered.

The original changes that this reverts were introduced in 3052137e.