vtkDataAssembly: simplify internal DOM
vtkDataAssembly now uses a DOM structure that directly matches the
assembly hierarchy i.e. use assembly node names as the names for nodes
in the DOM. This makes it easier to support complete set of XPath
selectors for selecting nodes.

This does add limitations to what names are supported as valid node
names in an assembly.