pio reader fixes and changes
Fixed a bug where sometimes a variable was determined to be a string
when it was actually numerical data.

Modified parsing of the .pio file. Comments can now be delimited with
'!' as well as '#'. Comments can be placed in the middle of a line, and
possibly on the same line as a command. Spaces are allowed at the
beginning of a line, as well as multiple spaces are allowed between
arguments.

Changed error messages to be clearer about what is wrong.