Add more OpenGL2 modules to Rendering group
The RenderingVolumeOpenGL2 module was in the Rendering group,
but not RenderingContextOpenGL2 or DomainsChemistryOpenGL2.
As a result, the latter two modules were not part of the
default build, even though they implement the backends for
modules that are part of the default build (DomainsChemistry,
RenderingContext2D, Charts, etc).

The fact that the VTK Python wheels provided modules without
backends has caused trouble for several users.

The RenderingOpenGL2 module itself was not part of the Rendering
group, either, but seems to have been brought in as a dependency
of RenderingVolumeOpenGL2.