For unstructured grids, enable point merging when requested
For vtkCutter, point merging is not an explicit option. It is enabled
by setting a vtkPointLocator that merges points. However, if a
vtkNonMergingPointLocator is set as the point locator, then points
will not be merged. Pass along the inferred merging option to
vtk3DLinearGridPlaneCutter when execution is deferred to it.