Fix TestVTKMImplicitDataArray
Previously, TestVTKMImplicitDataArray was not getting tested due to a
typo in the cmake file. Fixing the typo reveals an error in the code
which is fixed by this commit. The issue was that portals need to be
trivially copyable and std::function, which is not trivially copyable,
was being used in a portal for this test.