clang-tidy: fix some `performance-inefficient-string-concatenation` lints
The masking of the warning is left in place because some code ends up
reading way worse than it does right now. This commit basically gets the
easy and obvious improvements. If VTK ever gets access to a decent
formatting library, it can take care of the remaining patterns in use
(particularly in the shader builder code).