vtkCompositeDataSet: avoid std::accumulate over vtkCompositeDataSetRange
The `std::accumulate` implementation was causing invalid access when
copying state for the internals used by `CompositeDataSetRange`. It
appears as though the `end` iterator has some edge cases that are not
properly handled. This is filed as issue #18150.

Additionally, `vtkCompositeDataSetRange` is gaining some asserts to
catch such problems sooner when debugging.

This was discovered in `vtkCompositeDataSet::GetNumberOfElements` when
called via CTH AMR with MSVC.