Extend python deprecation warnings to constructors
This is for classes that are marked deprecated.  Creating an instance
of a deprecated class will now generate a DeprecationWarning.

Note that, for templated classes, the deprecation warning will use
the C++ name of the class e.g. vtkTypedArray<float> rather than the
Python name of the class, e.g. vtkTypedArray_IfE.  The wrappers use
ia64 mangling of the template parameters to produce a legal Python
identifier, but the original C++ identifier is easier on the eyes.