Several fixes for grid data types in ghost gen
* The ghost generator pipeline doesn't take anymore the input ghost
level information to learn how to peel off input ghosts
* `output->FastDelete()` removed, as this deleting has to be handled
elsewhere (in vtkGhostCellsGenerator)
* `FillReceivedGhosts` refactored to account for inputs not having cell
or point data. Ghost arrays are being filled and then the method is
aborted if there is no field data exchanged so far. There was also a
misshandling of ghost array copying: vtkFieldData doesn't allow to skip
arrays when setting tuples, so this is now done by hand.