Implementation change in vtkDIYUtilities::Link
The all_to_all communication in this method is unnecessary because when
the link map is being generated, it should be symmetric by construction,
so all the information is already present.

The signature of the method is also changed. It now uses polymorphism
when possible, and the `BlockT` template parameter is not relevant
anymore.

2 versions are written: one taking a std::set<int> as a parameter,
and the other taking a std::map<int, DummyT>. This gives a little
flexibility on the type of this link map, and allows to use an already
initialize map if the user needed to carry extra data when generating
the link map.

(cherry picked from commit 8400ec385e9ad167eb02be55e2de5855cded98c0)