Refactor XML Partitioned Dataset writer.
This refactors vtkXMLPartitionedDataSetWriter. This also lays the
groundwork to start create a parallel class hierarchy for all XML
writers.

Earlier implementation was producing incorrect results. Instead of
debugging through a complex piece of code, decided to refactor it and
start laying the core for a new way to write XML writers that doesn't
rely on deep class hierarchy.