Avoiding to interpolate several times the same point for Bezier surface
In vtkDataSetSurfaceFilter, nonlinear elements are triangulated.
For Bezier element, interpolation was performed for each corner of each triangle.
Since a triangle corner is often shared with other triangle corners, we are doing unnecessary interpolations.
The fix use the PointMap to check if a point has been already treated or not.