Provide a cache for the weights in GetInterpolatedPointId
For nonlinear elements with nonlinear subdivision > 1,
the surface is triangulated and GetInterpolatedPointId is called for each triangle corner.
This function was before allocating each time a vector for the weights.
We now move this allocation outside of this function,
 such that the allocation is done only once per unstructured grid.
So the weight is now part of that function argument list.
The old function has been marked as VTK_DEPRECATED_IN_9_1_0