vtkmodules: guard DLL directory addition from missing directories
This at least occurs in the wheel distribution where the install tree
does not look like the build tree because the wheel packaging rips a
directory out from within the build tree. Since the wheel is OK due to
the libraries being shipped right beside it, just check for the
existence of the DLL directories first.