Change in ghost buffer handling in ghost cell generator
The way ghosts were exchanged was a bit too rigid. It was relying on a
ghost buffer structure inside `BlockT` and using polymorphism, assuming
that we could stack geometry complexity (from implicit geometry to explicit
points to explicit cells). `vtkUnstructuredGrid` inner `vtkCellArray`
structure is too messy to do it by hand, and should be done using a
reader / writer, hence the impossibility to use polymorphism and add
layers of complexity in the exchanged data.

The ghost exchange is now done "a la carte" for each input type, calling
appropriate subroutines.

+ Renaming `ForGrids` to `ForStructuredData` as it is more correct