diy fix in vtkOverlappingCellsDetector
The way block linking was handled didn't quite scale up. Non symmetrical
link maps could be produced. The code has been cleaned and now relies on
the more robust / easy-to-use vtkDIYUtilities::Link to link relevant
blocks together.

The map was also storing `nullptr` ugs when blocks were not to be linked. It was
adding unnecessary complexity. Now, the filter doesn't allocate
anything, and relies on the existance or not of an ug for block at given
global id.