Wrapper hash table to track existing headers
This is to keep track of what header files exist on the system, in order
to potentially accelerate header file lookups.  The previous methodology,
of calling stat() for every full path created by appending the header
to every include directory, could be very slow on large projects where
the number of include directories was very large.

Co-authored-by: Andras Lasso <lasso@queensu.ca>