MPI fix for vtkProbeLineFilter
This filter was failing with MPI.
Internally, `vtkProbeFilter` is being used, and needs all ranks to have
the same probing input line, with the same samples.
In order to share those samples among ranks, the points on the line that
are being computed on each ranks are being shared between all ranks.
DIY is being used to do this communication. Hence, this filter is being
moved to Filters/ParallelDIY2

Adresses paraview/paraview#20769