Add InputArrayComponent member variable
Add a member variable that chooses which component in the active data
array to use for thresholding. If this value is set to a number larger
than or equal to the number of components in the input array to
process, the magnitude of each array tuple will be used for
thresholding instead of an individual component.

Updated test to exercise this feature and add some more correctness
checking of the filter's output by checking the output number of
points/cells in various scenarios.