vtkResourceFileLocator: use "found" and "not found" when searching
ParaView's test harness detects "failed" in the output as a failing
test, but this search is expected to have to go through a few search
paths in the general case. Avoid failing ParaView tests when resources
are not in the first place that they are looked for by using more
accurate wording about what is going on.