Fix shift+scale pausing.
The mapper' modification time was used to decide when
to re-upload VBOs, which meant that pausing and resuming
shift+scale computation would force a re-upload.
Furthermore, the composite polydata mapper holds multiple
instances of a mapper-helper that inherits the polydata
mapper and these instances must also have their computations
paused.