Revert "Merge topic 'backport-missed-mrs-9.0.3' into release"
This reverts commit 393370caf6fb397881f922ace363caee91db91e2, reversing
changes made to 7862ebd60c722f2fc52701afff8f8fc17ddc31dc.

While !6687 was requested to be backported (and done via !8134), it was
indicated that #18242 is an outstanding issue about something in the way
the backported commit changed event loop logic that causes VTK render
windows to not exit properly. Changes with such outstanding, unresolved
issues are not eligible for patch releases and as such is being
reverted.