Ghost cells generator: structured grid relative extent computation fixes
This applies for vtkImageData, vtkRectilinearGrid and vtkStructuredGrid

Relative extent was wrongly used when exchanging ghosts between 2
adjacent blocks. When extents are exchanged, the exchanged extent might
need to be shifted to match the discretization being used locally. For
example, vtkImageData with different Origin do not agrees on the spatial
signification of the extent. When this extent is calculated, it needs to
be used everywhere in the downstream to actually exchange ghost data.
This was not being done. This commit fixes this.