Add support for using both textures and interpolated scalars
When InterpolateScalars is enabled and the ColorMode is set to MapScalars,
vtkOpenGLPolyDataMapper internally uses a texture for scalar coloring.
Using this color texture should not prevent adding other textures with
multitexturing.
Besides supporting multitexturing, the actor texture defined with
actor->SetTexture(...) is also supported after explicitly specifying the
TCoords array on the mapper:
  mapper->MapDataArrayToMultiTextureAttribute("actortexture", ...)