Ghost cells generator: ghost point tagging fix for structured grids
The ghost point tagging had a bug with structured grids and was not
being tested. This commit fixes this bug and adds a test to ensure that
the ghost points will remain correctly tagged.

The bug was that points on the interface between the blocks were not
tagged as ghosts. One of the blocks needs to have them untagged, but the
other need them to be tagged as ghost points. The rule is now that the
block at the top - right - back to the other blocks "own" the point, and
all others tag those points as ghosts.