vtkGhostCellsGenerator: input ghost peeling optim for unstructured data
When an input poly data or unstructured grid has ghosts,
there used to be a deep copy of the
input into a temporary with its ghosts peeled off (calling
RemoveGhosts). This commit gets rid of this temporary and directly maps
the output cells and points with ghosts removed to the input cells and
points.