Ghost cells generator: cleaning legacy tests
Legacy tests were still in their former module. Now that they need
`vtkGhostCellsGenerator` rather than the previous implementation they
were using, we can move them to FiltersParallelDIY2, where
`vtkGhostCellsGenerator` lives.

In the way, I realized that a legacy test was not updated to use the new
backend. This is fixed.