Ghost cells generator: empty input partition bug fix
The case where an input partition is empty / virgin was being
mishandled for every input types. When the input is empty, the output
now directly just shallow-copies the input, and a pointer check is being
added for unstructured grid

The test has been updated to enforce inputs with empty partitions to
work.

Addresses paraview/paraview#20860