vtkFeatureEdges: Ghost handling improvements
Ghosts were wrongly handled in feature edges. Those cells were simply
skipped, but the presence of ghost cells messes up the topological
properties of edges that are part of a ghost cell. Now, when looking at
cells adjacent to an edge, the ghost cells are removed from the adjacent
cells