Skip color texture in multitexturing
6da590df added support for using textures with interpolate scalars.
However, the color texture used for scalar coloring was wrongly taken into
account for multitexturing, because the scalar color is already in
gl_FragData[0].
Fix this by excluding the color texture based on its name. We must update
the check to identify the first texture, because the color texture was
index 0.