GPUVolumeMapper: Clamp color values to avoid overflow/underflow
Color values as a result of complex lighting could overflow/underflow
because of the computations. This is a non-issue when directly rendering
to the framebuffer because they get clamped automatically based on the
type of the color buffer. However, in the case of ParaView/IceT
compositing, the raw values are extracted and used for compositing. This
makes the overflow/underflow issues apparent.