IO/Video: make `VTK_USE_VIDEO_FOR_WINDOWS_CAPTURE` an effective option
VTK should not silently ignore requested features if they don't work, so
instead add a cache variable for the class and error if it is detected
to not be available.

It is also renamed from `VTK_USE_VFW_CAPTURE` with compatibility of
using the old name's value as the new one's default if available.