Change change due to false positive from clang-tidy
The fix for a clang-tidy warning in a4d1bfc1 incorrectly invoked the
parent class method instead of the grandparent's method. This led to a
slew of error messages in vtkCompositeControlPointsItem whenever the
number of control points in the opacity function exceeded the number
of control points in the color transfer function because the parent
class's implementation of DrawPoint makes assumptions that do not
hold, hence the invocation of the grandparent class's method.  Fix by
invoking the grandparent's method as before the clang-tidy fix, and
tell tidy to ignore it.