Fix Python tuple to non-ref std::vector conversion
When calling a method with a "std::vector&" parameter, the called
method might modify the vector, so a mutable sequence (eg. a list)
must be passed from Python.

Unfortunately, this same restriction (of requiring a mutable rather
than non-mutable sequence) was also being placed on "std::vector"
parameters that were not passed by reference.  This patch removes
the resriction.