Fix expansion of stl template args in wrapping
When wrapped methods have parameter types like std::vector<T>,
we check if "T" is a typedef.  Unfortunately, the check code
did not work for template args like std::array<T,4>.

This new code verifies that the template arg can be parsed as a
type before assuming it is a type.

The code is still unable to properly handle template IDs like
std::array<T,N> where 'N' is a name that resolves to a constant,
because the hierarchy files only store type names (for now).