quadraticIntersection: Render the whole render window
During the implementation of vtkTriQuadraticPyramid, a new line was added
and therefore some pixels had garbage values, because no renderer was handling
the rendering of the empty areas. This commit renders by default the whole
render window to deal with the garbage values. Additionally, the resolution
of the image has been increased to capture changes that should have been
captured when the IntersectWithLine function of vtkTriQuadraticPyramid was fixed.