vtkExtractGrid: fix issues with distributed execution
vtkExtractGrid, when executing in a distributed environment, ended up
updating the internal helper to execute on subextent but failed to
restore it. This was causing `RequestUpdateExtent` to keep executing on
a shrinking whole extent raising warnings (which in reality should have
reported as errors).

This fixes that issue. Also added a test to confirm that these errors
don't pop up.