vtkContour3DLinearGrid: add ComputeScalars option
Add option ComputeScalars. When enabled, and when the
InterpolateAttributes option is on, vtkContour3DLinearGrid will add an
array corresponding to the array used to compute the contour and
populate it with values. This option brings this filter closer to the
full functionality of vtkContourFilter. This option is off by default.