Ghost cells generator: kdtree point locator
In the unstructured data subroutine, a `vtkStaticPointLocator` was used
to match interfacing points with points sent by neighbors. In the
instance of unstructured grids with a lot of discrepencies in point
density over the data set, this point locator could become extremely
slow. This is because the static point locator uses cubic buckets and
brute-force search inside the bucket in which the query point is. In the
instance of querying in a bucket with high resolution data, the search
is very slow.

To speed up the filter for this case scenario, `vtkStaticPointLocator`
is replaced by `vtkKdTreePointLocator`.

Addresses paraview/paraview#20926