Ghost cells generator: bug fix when input ghosts are present
* In the presence of ghost cells in the input, the ghost cells generator
generates maps so one can know the point id in the input given a point
in the output. This remapping was missing for the points at the
interface between 2 adjacent partitions.

* The naming of the cell id / point id map was wrong. `InputToOutput`
should have been named `OutputToInput` as we are fetching points from
the input using the id in the output. An instance of `InpputToOutput`
was however necessary for point ids, so it also has been added.

* A potential bug fix has been resolved. This should not have been a
problem as if all the points to be peeled off in the input are at the
end of the point array, then the bug was not triggered. A point
remapping from the input to the output was needed when deep copying
input cells into the output. `InputToOutputPointIdRedirectionMap` is
used for this case.

* Input unstructured grids could have a non nullptr `Faces` and
`FaceLocations` empty array. The filter was only checking for `nullptr`
on those. A check on the number of values has beend added

* Testing is being update to check for distorted cells in the output.

Addresses paraview/paraview#20906

(cherry picked from commit f8412719eed8713ba95c48e5e7f27eda4b123902)