Fix a loadtime issue with vtkOpenGLTexture
This MR fixes an issue with vtkOpenGLTexture where it was not updating
it's load time when using an external texture object. This would cause
issues when switching between render windows for shared texture objects.
The problem could lead to a crash when one of the windows sharing a
texture is freed.